Add server route to get all known balances
