+++ src/errors.rs
@@ -41,6 +41,9 @@ pub enum EndpointError {
     #[fail(display = "unable to get relay status.")]
     UnableToGetStatus,
 
+    #[fail(display = "unable to get nectar balances.")]
+    UnableToGetBalances,
+
     #[fail(display = "timeout")]
     Timeout,
 }
@@ -52,6 +55,7 @@ impl ResponseError for EndpointError {
             EndpointError::BadTransactionHash(_) => HttpResponse::new(http::StatusCode::BAD_REQUEST),
             EndpointError::UnableToSend => HttpResponse::new(http::StatusCode::INTERNAL_SERVER_ERROR),
             EndpointError::UnableToGetStatus => HttpResponse::new(http::StatusCode::INTERNAL_SERVER_ERROR),
+            EndpointError::UnableToGetBalances => HttpResponse::new(http::StatusCode::INTERNAL_SERVER_ERROR),
             EndpointError::Timeout => HttpResponse::new(http::StatusCode::REQUEST_TIMEOUT),
         }
     }
