+++ src/server/mod.rs
@@ -1,6 +1,10 @@
 pub mod endpoint;
+pub mod handler;
 
 use super::errors;
 use super::eth;
 use super::relay;
 use super::transfers;
+
+pub use self::endpoint::RequestType;
+pub use self::handler::HandleRequests;
