+++ src/transfers/past.rs
@@ -4,7 +4,7 @@ use web3::futures::future;
 use web3::futures::prelude::*;
 use web3::futures::sync::mpsc;
 use web3::futures::try_ready;
-use web3::types::{Address, BlockNumber, FilterBuilder, U256};
+use web3::types::{Address, BlockNumber, FilterBuilder, TransactionReceipt, H256, U256};
 use web3::{DuplexTransport, ErrorKind};
 
 use super::eth::contracts::TRANSFER_EVENT_SIGNATURE;
@@ -267,5 +267,132 @@ impl<T: DuplexTransport + 'static> Future for ValidateAndApproveTransfer<T> {
             handle.spawn(self.transfer.approve_withdrawal(&target));
         }
         Ok(Async::Ready(()))
+    }
+}
+
+pub enum FindTransferState {
+    ExtractTransfers(Box<Future<Item = Vec<Transfer>, Error = ()>>),
+    FetchReceipt(Box<Future<Item = Option<TransactionReceipt>, Error = ()>>),
+}
+
+/// Future to find a vec of transfers at a specific transaction
+pub struct FindTransferInTransaction<T: DuplexTransport + 'static> {
+    hash: H256,
+    source: Rc<Network<T>>,
+    state: FindTransferState,
+}
+
+impl<T: DuplexTransport + 'static> FindTransferInTransaction<T> {
+    /// Returns a newly created FindTransferInTransaction Future
+    ///
+    /// # Arguments
+    ///
+    /// * `source` - Network where the transaction took place
+    /// * `hash` - Transaction hash to check
+    pub fn new(source: &Rc<Network<T>>, hash: &H256) -> Self {
+        let web3 = source.web3.clone();
+        let network_type = source.network_type;
+        let future = web3.clone().eth().transaction_receipt(*hash).map_err(move |e| {
+            error!("error getting transaction receipt on {:?}: {:?}", network_type, e);
+        });
+        let state = FindTransferState::FetchReceipt(Box::new(future));
+        FindTransferInTransaction {
+            hash: *hash,
+            source: source.clone(),
+            state,
+        }
+    }
+}
+
+impl<T: DuplexTransport + 'static> Future for FindTransferInTransaction<T> {
+    type Item = Vec<Transfer>;
+    type Error = ();
+    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
+        loop {
+            let source = self.source.clone();
+            let network_type = source.network_type;
+            let hash = self.hash;
+            let next = match self.state {
+                FindTransferState::ExtractTransfers(ref mut future) => {
+                    let transfers = try_ready!(future.poll());
+                    if transfers.is_empty() {
+                        warn!("no relay transactions found on {:?} at {:?}", network_type, hash);
+                        return Err(());
+                    } else {
+                        return Ok(Async::Ready(transfers));
+                    }
+                }
+                FindTransferState::FetchReceipt(ref mut future) => {
+                    let receipt = try_ready!(future.poll());
+                    match receipt {
+                        Some(r) => {
+                            if r.block_hash.is_none() {
+                                error!("receipt did not have block hash on {:?}", network_type);
+                                return Err(());
+                            }
+                            let block_hash = r.block_hash.unwrap();
+                            r.block_number.map_or_else(
+                                || {
+                                    error!("receipt did not have block number on {:?}", network_type);
+                                    Err(())
+                                },
+                                |receipt_block| {
+                                    let future = source
+                                        .web3
+                                        .clone()
+                                        .eth()
+                                        .block_number()
+                                        .and_then(move |block| {
+                                            let mut transfers = Vec::new();
+                                            if let Some(confirmed) =
+                                                block.checked_rem(receipt_block).map(|u| u.as_u64())
+                                            {
+                                                if confirmed > source.confirmations {
+                                                    let logs = r.logs;
+                                                    for log in logs {
+                                                        info!(
+                                                            "found log at {:?} on {:?}: {:?}",
+                                                            hash, network_type, log
+                                                        );
+                                                        if log.topics[0] == TRANSFER_EVENT_SIGNATURE.into()
+                                                            && log.topics[2] == source.relay.address().into()
+                                                        {
+                                                            let destination: Address = log.topics[1].into();
+                                                            let amount: U256 = log.data.0[..32].into();
+                                                            if destination == Address::zero() {
+                                                                info!("found mint on {:?}. Skipping", network_type);
+                                                                continue;
+                                                            }
+                                                            let transfer = Transfer {
+                                                                destination,
+                                                                amount,
+                                                                tx_hash: hash,
+                                                                block_hash,
+                                                                block_number: receipt_block,
+                                                                removed: false,
+                                                            };
+                                                            transfers.push(transfer);
+                                                        }
+                                                    }
+                                                }
+                                            }
+                                            Ok(transfers)
+                                        })
+                                        .map_err(move |e| {
+                                            error!("error getting current block on {:?}: {:?}", network_type, e);
+                                        });
+                                    Ok(FindTransferState::ExtractTransfers(Box::new(future)))
+                                },
+                            )
+                        }
+                        None => {
+                            error!("unable to find {:?} on {:?}", hash, source.network_type);
+                            return Err(());
+                        }
+                    }?
+                }
+            };
+            self.state = next;
+        }
     }
 }
